pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ml-curso"
version = "0.1.0"
description = "Repositório Guiado de Machine Learning com Python"
requires-python = ">=3.10,<3.13"
dependencies = [
    # Core dependencies
    "numpy>=1.25.0,<2.0",
    "pandas>=2.1.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.2",
    "seaborn>=0.12.2",
    "scipy>=1.11.1",
    
    # Deep Learning
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    
    # Notebook execution
    "jupyter>=1.0.0",
    "papermill>=2.4.0",
    "nbclient>=0.8.0",
    "ipykernel>=6.25.0",
    
    # Data handling
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
    
    # Utilities
    "tqdm>=4.65.0",
    "click>=8.1.6",
    
    # Data visualization and missing data analysis
    "missingno>=0.5.2",
]

[project.optional-dependencies]
dev = [
    # Development tools
    "ruff==0.0.280",
    "black==23.7.0",
    "isort==5.12.0",
    "mypy==1.5.1",
    "pytest==7.4.0",
    "pytest-cov==4.1.0",
    
    # Type stubs
    "types-PyYAML>=6.0.12",
    "types-tqdm>=4.66.0",
]

[project.scripts]
ml-curso = "scripts.tasks:main"

[tool.setuptools.packages.find]
include = ["core*", "scripts*"]
exclude = ["modules*", "datasets*", "tests*", "docs*"]

[tool.uv]
dev-dependencies = [
    "ruff==0.0.280",
    "black==23.7.0",
    "isort==5.12.0",
    "mypy==1.5.1",
    "pytest==7.4.0",
    "pytest-cov==4.1.0",
    
    # Type stubs
    "types-PyYAML>=6.0.12",
    "types-tqdm>=4.66.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"modules/**/exercises/*.ipynb" = ["F401", "F841"]  # Allow unused imports/variables in exercises

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
